+++ pinterest_embed_plugin/lib/src/scripts/pinterest_script_loader.dart
@@ -1,10 +1,18 @@
 // ignore_for_file: avoid_web_libraries_in_flutter
-import 'package:flutter/widgets.dart';
+import 'dart:async';
+
+import 'package:pinterest_embed_plugin/src/scripts/js_interop/pin_utils_interop.dart';
+import 'package:pinterest_embed_plugin/src/utils/hydration_state.dart';
 import 'package:web/web.dart' as web;
 
 class PinterestScriptLoader {
   static const String _scriptId = 'pinterest-pinit-js';
   static const String _scriptSrc = 'https://assets.pinterest.com/js/pinit.js';
+
+  static final StreamController<HydrationState> _controller =
+      StreamController.broadcast();
+
+  static Stream<HydrationState> get hydrationStream => _controller.stream;
 
   static void ensureScriptAndHydrate() {
     final existing = web.document.getElementById(_scriptId);
@@ -19,6 +27,7 @@ class PinterestScriptLoader {
 
       // When script finishes loading, hydrate embeds
       script.onLoad.listen((_) {
+        _controller.add(HydrationState.loading);
         _tryHydrate();
       });
 
@@ -31,19 +40,31 @@ class PinterestScriptLoader {
     }
   }
 
-  static void _tryHydrate({int attempts = 5}) {
-    if (attempts <= 0) return;
+  static void _tryHydrate({int attempts = 7, int delayMs = 400}) {
+    if (attempts <= 0) {
+      _controller.add(HydrationState.error);
+      return;
+    }
 
     try {
-      (web.window as dynamic).PinUtils?.build();
-      debugPrint("[Pinterest] Hydration succeeded.");
-      print("[Pinterest] Hydration succeeded. attempt $attempts");
+      final pinUtils = getPinUtils();
+
+      if (pinUtils != null) {
+        
+        pinUtils.build();
+
+        _controller.add(HydrationState.success);
+        return;
+      } else {
+        throw Exception("PinUtils is not yet ready.");
+      }
     } catch (err) {
+      print(
+        "[Pinterest] Hydration failed. attempt $attempts,\nError: $err \nretry Delay $delayMs",
+      );
 
-      print("[Pinterest] Hydration failed. attempt $attempts,\nErro: $err");
-
-      Future.delayed(const Duration(milliseconds: 300), () {
-        _tryHydrate(attempts: attempts - 1);
+      Future.delayed(Duration(milliseconds: delayMs), () {
+        _tryHydrate(attempts: attempts - 1, delayMs: (delayMs * 1.5).toInt());
       });
     }
   }