+++ pinterest_embed_plugin/lib/src/utils/hydration_state.dart
@@ -1,0 +1,5 @@
+enum HydrationState {
+  loading,
+  success, 
+  error 
+}