+++ pinterest_embed_plugin/lib/src/widgets/loader_widget.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class LoaderWidget extends StatelessWidget {
+  const LoaderWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        const CircularProgressIndicator(),
+        const SizedBox(height: 12),
+        Text(
+          "Preparing embeds",
+          style: GoogleFonts.montserrat(
+            fontWeight: FontWeight.bold,
+            color: Colors.black,
+            fontSize: 20,
+          ),
+          textAlign: TextAlign.center,
+        ),
+        const SizedBox(height: 6),
+        Text(
+          "We are getting your pinterest embed widget set up. \nPlease wait",
+          style: GoogleFonts.inter(
+            fontWeight: FontWeight.normal,
+            color: Colors.grey,
+            fontSize: 16,
+          ),
+          textAlign: TextAlign.center,
+        ),
+        const SizedBox(height: 6),
+      ],
+    );
+  }
+}